refactor: applications structure wip
