+++ backend/src/modules/application/constants/application.constants.ts
@@ -6,7 +6,8 @@ export const ORGANIZATION_ALL_APPS_COLUMNS = [
   'application.short_description as "shortDescription"',
   'application.login_link as "loginLink"',
   'application.website as website',
-  'ongApp.status as status',
+  'application.status as status',
+  'ongApp.status as "ongStatus"',
   'ongApp.created_on as "createdOn"',
   'application.status as "applicationStatus"',
   'application.type as type',
