+++ backend/src/modules/application/interfaces/application-with-ong-status.interface.ts
@@ -2,6 +2,7 @@ import { ApplicationPullingType } from '../enums/application-pulling-type.enum';
 import { ApplicationStatus } from '../enums/application-status.enum';
 import { ApplicationTypeEnum } from '../enums/ApplicationType.enum';
 import { OngApplicationStatus } from '../enums/ong-application-status.enum';
+import { UserOngApplicationStatus } from '../enums/user-ong-application-status.enum';
 
 export interface ApplicationWithOngStatus {
   id: number;
@@ -10,7 +11,8 @@ export interface ApplicationWithOngStatus {
   shortDescription: string;
   loginLink: string;
   website: string;
-  status: OngApplicationStatus | ApplicationStatus;
+  status: ApplicationStatus;
+  ongStatus: OngApplicationStatus;
   type: ApplicationTypeEnum;
 }
 
@@ -19,5 +21,6 @@ export interface ApplicationWithOngStatusDetails
   steps: string[];
   description: string;
   videoLink: string;
+  userStatus: UserOngApplicationStatus;
   pullingType: ApplicationPullingType;
 }
