+++ frontend/src/pages/requests/interfaces/OngApplication.interface.ts
@@ -5,7 +5,6 @@ export enum OngApplicationStatus {
   ACTIVE = 'active',
   PENDING = 'pending',
   RESTRICTED = 'restricted',
-  DISABLED = 'disabled',
   PENDING_REMOVAL = 'pending-removal',
 }
